impl: move language/common.go to genclient/
This removes a dependency from package parser on package language.
