+++ generator/internal/parser/openapi/openapi.go
@@ -23,7 +23,6 @@ import (
 	"strings"
 
 	"github.com/googleapis/google-cloud-rust/generator/internal/genclient"
-	"github.com/googleapis/google-cloud-rust/generator/internal/language"
 	"github.com/pb33f/libopenapi"
 	"github.com/pb33f/libopenapi/datamodel/high/base"
 	v3 "github.com/pb33f/libopenapi/datamodel/high/v3"
@@ -97,9 +96,9 @@ func makeAPI(serviceConfig *serviceconfig.Service, model *libopenapi.DocumentMod
 	packageName := ""
 	if serviceConfig != nil {
 		for _, api := range serviceConfig.Apis {
-			packageName, serviceName = language.SplitApiName(api.Name)
+			packageName, serviceName = genclient.SplitApiName(api.Name)
 			// Keep searching after well-known mixin services.
-			if !language.WellKnownMixin(api.Name) {
+			if !genclient.WellKnownMixin(api.Name) {
 				break
 			}
 		}
